#58 fixed token generation

+++ devcom_mvp/src/main/java/com/devcom/community/Community.java
@@ -21,7 +21,7 @@ public class Community {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer communityId;
+    private Long communityId;
     @Column(nullable = false)
     private String name;
     private String description;
