+++ devcom_mvp/src/main/java/com/devcom/community/CommunityRepository.java
@@ -2,5 +2,5 @@ package com.devcom.community;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CommunityRepository extends JpaRepository<Community, Integer> {
+public interface CommunityRepository extends JpaRepository<Community, Long> {
 }
